requirements.txt
torch
transformers
accelerate
Pillow
pandas
numpy
tqdm
wandb
scipy
sklearn
